fifteen/random_test.go
// fifteen - 15-puzzle solver.
// Copyright (C) 2019  Tulir Asokan
//
// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU Affero General Public License as published by
// the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU Affero General Public License for more details.
//
// You should have received a copy of the GNU Affero General Public License
// along with this program.  If not, see <https://www.gnu.org/licenses/>.

package fifteen

import (
	"github.com/stretchr/testify/assert"
	"testing"
)

func TestPuzzle_Shuffle(t *testing.T) {
	puzzle := NewSolvedPuzzle(4)
	assert.True(t, puzzle.Solvable())
	assert.True(t, puzzle.IsSolved())
	puzzle.Shuffle(200)
	assert.True(t, puzzle.Solvable())
	assert.False(t, puzzle.IsSolved())
}